removed BaseSequence
