+++ mod/config.cpp
@@ -2,7 +2,6 @@
 
 class CfgPatches {
     class ZT {
-        //units[] = {"ModuleCheckVIP"};
         requiredVersion = 1.0;
         requiredAddons[] = {"A3_Modules_F"};
     };
